implemented blogs

+++ App/Screens/ProfileTab/BlogScreen.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useCallback, useState } from 'react';
-import { TextInput, Button, StyleSheet, Text, View, Image, FlatList, Dimensions, TouchableOpacity } from 'react-native';
+import { TextInput, Button, StyleSheet, Text, View, Image, FlatList, Dimensions, TouchableOpacity, Alert } from 'react-native';
 
 import { useSelector, useDispatch } from 'react-redux';
 
@@ -15,8 +15,6 @@ const BlogScreen = (props) => {
 
 
     const images = props.route.params?.blog.images;
-    const numComments = props.route.params?.blog.numComments ?? 0
-    const numReacts = props.route.params?.blog.numReacts ?? 0
     const blog = props.route.params?.blog
 
     const [error, setError] = useState('')
@@ -33,7 +31,7 @@ const BlogScreen = (props) => {
     const [change, setChange] = useState(0);    //this forces comments to re render on each touch
     // console.log(comments)
 
-    const loadUserBlogComments = useCallback(async (blogId) => {
+    const loadUserBlogComments = useCallback(async () => {
         try {
             await dispatch(magazineActions.fetchUserBlogComments(blogId))
         }
@@ -42,7 +40,7 @@ const BlogScreen = (props) => {
         }
     }, [blogId])
 
-    const loadUserBlogReacts = useCallback(async (blogId) => {
+    const loadUserBlogReacts = useCallback(async () => {
         try {
             await dispatch(magazineActions.fetchUserBlogReacts(blogId))
 
@@ -50,71 +48,80 @@ const BlogScreen = (props) => {
         catch (err) {
             console.log(err)
         }
-    })
+    }, [blogId])
 
-    const reactUserBlog = useCallback(async (blogId) => {
+    const reactUserBlog = useCallback(async () => {
         try {
             await dispatch(magazineActions.reactUserBlog(blogId))
-            setError('')
+
+            setChange(state => state + 1)
         }
         catch (err) {
-            setError(err.message)
+
             console.log(err);
         }
-    })
-    const unReactUserBlog = useCallback(async (blogId) => {
+    }, [blogId])
+    const unReactUserBlog = useCallback(async () => {
         try {
             await dispatch(magazineActions.unReactUserBlog(blogId))
-            setError('')
+
+            setChange(state => state - 1)
+
         }
         catch (err) {
-            setError(err.message)
+
             console.log(err);
         }
-    })
+    }, [blogId])
 
-    const commentUserBlog = useCallback(async (blogId, comment) => {
+    const commentUserBlog = useCallback(async (comment) => {
         try {
             await dispatch(magazineActions.commentUserBlog(blogId, comment))
-            setError('')
+
         }
         catch (err) {
-            setError(err.message)
+
             console.log(err);
         }
-    })
-    const deleteCommentUserBlog = useCallback(async (commentId, postId) => {
+    }, [blogId])
+    const deleteCommentUserBlog = useCallback(async (commentId) => {
         try {
             await dispatch(magazineActions.deleteCommentUserBlog(commentId))
-            setError('')
+
         }
         catch (err) {
-            setError(err.message)
+
             console.log(err);
         }
     })
 
-    const deleteBlog = useCallback(async (blogId) => {
+    const deleteBlog = useCallback(async () => {
         try {
-            await dispatch(profileActions.deleteUserBlog(blogId))
-            setError('')
+
+            Alert.alert('Delete Blog?', "Are you sure you want to delete this blog?", [
+                {
+                    text: "Delete",
+                    onPress: async () => {
+                        await dispatch(profileActions.deleteUserBlog(blogId))
+                        await dispatch(profileActions.fetchMyBlogs())
+                        props.navigation.goBack();
+
+                    }
+                },
+                {
+                    text: 'Cancel',
+
+                    style: 'cancel'
+                }
+            ])
+
+
         }
         catch (err) {
-            setError(err.message)
+
             console.log(err);
         }
-    })
-
-
-
-    const renderImage = (itemData) => {
-        // console.log(itemData.item.image)
-        return (
-
-            <Image source={itemData.item.image} style={styles.postImage}
-                resizeMode="contain" />
-        )
-    }
+    }, [blogId])
 
     useEffect(() => {
 
@@ -122,106 +129,114 @@ const BlogScreen = (props) => {
 
     }, [blogId])
 
+
+
+
     const renderComment = (itemData) => {
         if (itemData.index === 0) {
             return (
-                <>
-                    <FlatList
+                <View style={styles.blogBody}>
+                    <Text style={styles.title}>Title here</Text>
+
+                    <TouchableOpacity>
+                        <View style={styles.nameDP}>
+
+                            {/* <Image style={styles.DPImage} source={post?.logo ?? logo} /> */}
+                            <View style={styles.nameContainer}>
+                                <Text style={styles.Name}> {blog.name} </Text>
+                                <Text style={styles.username}> {blog.username} . {blog.date} </Text>
+                            </View>
+
+
+                        </View>
+                    </TouchableOpacity>
+
+
+                    {/* <FlatList
                         data={images}
                         pagingEnabled={true}
                         horizontal={true}
                         renderItem={renderImage}
-                    />
+                    /> */}
 
-                    <FlatList data={blog.texts} renderItem={renderText} />
+                    <Image source={images[0]?.image} style={styles.image}
+                        resizeMode="contain" />
 
 
+                    {/* <FlatList
 
-                    <View style={styles.nums}>
-                        <TouchableOpacity onPress={() => {
-                            loadUserBlogReacts(blogId);
-                            setShowReacts(state => !state)
-                            // console.log(postComments)
-                        }}>
-                            <Text>{numReacts} Reacts</Text>
-                        </TouchableOpacity>
+                        data={blog.texts}
+                        renderItem={renderText}
+                    /> */}
 
-                        <TouchableOpacity onPress={() => {
+                    <Text style={styles.text}>{blog.texts[0].text}</Text>
 
-                            loadUserBlogComments(blogId)
-                            setChange(state => state + 1)
-                        }}>
-                            <Text>{numComments} Comments</Text>
-                        </TouchableOpacity>
+                    <Text style={styles.text}>{blog.texts[1].text}</Text>
+                    <Text style={styles.text}>{blog.texts[2].text}</Text>
 
-                    </View>
-
-                    {showReacts ? <FlatList listKey={blogId + "1"} data={reacts} renderItem={renderReact} ListEmptyComponent={() => (
-                        <Text>no reacts</Text>
-                    )} /> : null}
-                </>
+                </View>
             )
         }
 
         if (itemData.index === comments.length + 1) {
             return (
                 <View>
-                    <View style={styles.LikeComment}>
+
+
+                    <View style={styles.reactsCommentsContainer}>
                         <TouchableOpacity style={styles.Like} onPress={async () => {
 
 
                             if (blog.hasReacted === 1) {
                                 //unlike
                                 await unReactUserBlog(blog.id)
-                                if (error === '') {
 
 
-                                    // flatListRef.current.recordInteraction()
-                                    blog.hasReacted = 0;
-                                    blog.numReacts -= 1;
-                                    setChange(state => state - 1)
-                                }
+
+                                // flatListRef.current.recordInteraction()
+                                blog.hasReacted = 0;
+                                blog.numReacts -= 1;
+
+
 
                             }
                             else {
-                                await reactUserBlog(blog.id)
-                                if (error === '') {
-                                    blog.hasReacted = 1;
-                                    blog.numReacts += 1;
-                                    setChange(state => state + 1)
-                                }
+                                reactUserBlog(blog.id)
+
+                                blog.hasReacted = 1;
+                                blog.numReacts += 1;
+
+
 
 
                             }
-
-
-
-
                         }}>
 
-                            {blog.hasReacted === 1 ? <AntDesign name="like1" size={40} color='black' /> : <AntDesign name="like2" size={40} color='black' />}
+
+                            {blog.hasReacted === 1 ? <MaterialCommunityIcons name="heart-multiple" size={30} color='#E1306C' /> : <MaterialCommunityIcons name="heart-multiple-outline" size={30} color='black' />}
+
+                            <Text style={styles.number}>{blog.numReacts}</Text>
+
 
                         </TouchableOpacity>
 
-                        <View style={styles.commentContainer}>
-                            <TextInput placeholder="Comment" style={styles.Comment} onChangeText={setComment} />
-                            <TouchableOpacity style={styles.sendComment}
-                                onPress={() => {
-                                    comment !== '' ?
-                                        commentUserBlog(blog.id, comment) : null;
-                                    setChange(state => state + 1)
 
-                                }}
 
-                            >
-                                <Ionicons name="md-send" size={30} />
-                            </TouchableOpacity>
+
+                        <View
+                            style={styles.commentIcon}
+
+                        >
+                            <MaterialCommunityIcons name="comment-multiple" color="black" size={30} />
+                            <Text style={styles.number}>{blog.numComments}</Text>
                         </View>
 
 
 
-
                     </View>
+
+
+
                     {blog.userId === userId ? <Button title="delete blog" onPress={() => {
                         console.log(blog.id)
 
@@ -235,45 +250,18 @@ const BlogScreen = (props) => {
 
         return (
             <View style={styles.comment}>
-                <Text>{itemData.item.username}:  </Text>
-                <Text>{itemData.item.comment}</Text>
-                {itemData.item.commenterId === userId ?
-                    (
-                        <TouchableOpacity
-                            onPress={() => {
-                                deleteCommentUserBlog(itemData.item.id)
-                                setChange(state => state + 1)
+                <View style={styles.commentUsernameContainer} >
+                    <Text style={styles.commentUsername}>{itemData.item.username} .  </Text>
+                    <Text style={styles.commentDate}>{itemData.item.date}</Text>
 
-                            }}
-                            style={styles.commentX}
+                </View>
 
-                        >
-                            <Text>X</Text>
-                        </TouchableOpacity>
-                    ) : null}
+                <Text style={styles.commentText}>{itemData.item.comment}</Text>
             </View>
         )
 
 
     }
-
-    const renderReact = (itemData) => {
-        return (
-            <Text>{itemData.item.username}:   {itemData.item.type}</Text>
-        )
-    }
-
-    const renderText = (itemData) => {
-        return (
-            <View>
-                <Text>Text {itemData.index} </Text>
-                <Text>{itemData.item.text}</Text>
-            </View>
-
-
-        )
-    }
-
 
     return (
         <View>
@@ -295,51 +283,115 @@ const BlogScreen = (props) => {
 export default BlogScreen;
 
 const styles = StyleSheet.create({
-    postGridItem: {
-        margin: 10,
-        width: Dimensions.get('window').width / 3.5
+
+    blogBody: {
+        backgroundColor: 'white',
+        marginVertical: 20
+
     },
-    postImage: {
+    title: {
+        fontSize: 30,
+        fontFamily: 'serif',
+        fontWeight: '600',
+        textAlign: 'center',
+        marginVertical: 20
+    },
+    nameDP: {
+        flexDirection: 'row',
+        paddingHorizontal: 10
+    },
+
+    DPImage:
+    {
+        width: 40,
+        height: 40,
+        borderRadius: 20
+    },
+    nameContainer: {
+        marginLeft: 10,
+        flexDirection: "column"
+
+    },
+    Name:
+    {
+        fontWeight: '700',
+        fontSize: 18
+    },
+    username: {
+        fontSize: 15,
+        color: 'grey'
+
+    },
+
+    image: {
         height: Dimensions.get('window').width,
         width: Dimensions.get('window').width,
     },
-    nums: {
-        width: '100%',
-        alignItems: 'center'
+    text: {
+        fontSize: 18,
+        marginHorizontal: 10,
+        marginVertical: 20
     },
-    LikeComment:
-    {
-        paddingTop: 15,
-        flexDirection: 'row'
+    reactsCommentsContainer: {
+        flexDirection: 'row',
+        width: '100%',
+        alignItems: 'center',
+        justifyContent: 'space-between',
+        paddingHorizontal: 10,
+        marginVertical: 20
+
+
     },
     Like:
     {
         paddingRight: 15,
         paddingLeft: 10,
-        flex: 1
+        flexDirection: 'row',
+        alignItems: 'center',
+        width: 70
     },
-    Comment:
-    {
-        flex: 3,
-        paddingLeft: 5,
-        height: 40,
-        borderColor: 'black',
-        backgroundColor: 'grey'
+    number: {
+        fontSize: 15,
+        fontWeight: '700',
+        color: 'grey',
+        marginLeft: 2,
+        flex: 1,
+
+
+    },
+    commentIcon: {
+        flexDirection: 'row',
+        width: 70,
+        justifyContent: 'flex-end',
+        alignItems: 'center',
     },
     comment: {
-        flexDirection: 'row',
-        height: 100,
-        margin: 10
+        flexDirection: 'column',
+        backgroundColor: 'white',
+        paddingHorizontal: 10,
+        marginVertical: 10,
+        paddingVertical: 5
+
     },
-    commentX: {
-        marginLeft: 30
+    commentUsernameContainer: {
+        flexDirection: 'row'
+
+    },
+    commentUsername: {
+        fontWeight: '700',
+        fontSize: 15
+
+    },
+    commentDate: {
+        fontWeight: '600',
+        color: 'grey'
+
+    },
+    commentText: {
+        fontSize: 18,
+        marginLeft: 10
+
     },
 
-    commentContainer: {
-        flexDirection: 'row',
-        flex: 1
-    },
-    sendComment: {
-        marginLeft: 3
-    }
+
 })
